Fold conductor_api into core
