--- crates/conductor_api/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "sx_conductor_api"
-version = "0.0.1"
-description = "holochain conductor api"
-license = "GPL-3.0-only"
-homepage = "https://github.com/holochain/holochain-rust"
-documentation = "https://github.com/holochain/holochain-rust"
-authors = ["Holochain Core Dev Team <devcore@holochain.org>"]
-edition = "2018"
-
-[dependencies]
-async-trait = "0.1.22"
-sx_types = { path = "../types" }
-thiserror = "1.0.10"
-tokio = { version = "0.2.11", features = ["macros", "rt-threaded", "rt-util", "sync"] }
-log = "0.4.8"
-mockall = "0.6.0"
